Add files via upload
This is a very basic and simple Music Player. It can do several functions like playing next, previous or current song again, then you can also delete a song, or you can print the list of the song. Basic knowledge of OOPs concept and Collections Framework will be required to understand this code and to create on your own.